# Optimize comparisons against Guid.Empty