--- a/src/coreclr/jit/importer.cpp
+++ b/src/coreclr/jit/importer.cpp
@@ -4210,9 +4210,9 @@ GenTree* Compiler::impImportStaticReadOnlyField(CORINFO_FIELD_HANDLE field, CORI
                 JITDUMP("success! Optimizing to ASG(struct, 0).");
                 unsigned structTempNum = lvaGrabTemp(true DEBUGARG("folding static ro fld empty struct"));
                 lvaSetStruct(structTempNum, fieldClsHnd, false);
-                GenTreeLclVar* tree = impCreateLocalNode(structTempNum DEBUGARG(0));
+                GenTreeLclVar* tree = gtNewLclvNode(structTempNum, TYP_STRUCT);
                 impAppendTree(gtNewBlkOpNode(tree, gtNewIconNode(0), false, false), CHECK_SPILL_NONE, impCurStmtDI);
-                return gtClone(tree); // clonning GenTreeLclVar node
+                return gtNewLclvNode(structTempNum, TYP_STRUCT);
             }
 
             JITDUMP("getReadonlyStaticFieldValue returned false - bail out.");