@EgorBo 2022-10-23
```csharp
static decimal Test() => decimal.Zero;
```
Codegendiff:
```diff
; Assembly listing for method Program:Test():System.Decimal
       C5F877               vzeroupper 
-      48B86000C07FC0020000 mov      rax, 0x2C07FC00060
-      488B00               mov      rax, gword ptr [rax]
-      C5F9104008           vmovupd  xmm0, xmmword ptr [rax+08H]
+      C5F857C0             vxorps   xmm0, xmm0
       C5F91101             vmovupd  xmmword ptr [rcx], xmm0
       488BC1               mov      rax, rcx
       C3                   ret      
; Total bytes of code 29
```